+++ roseau/load_flow/utils/__init__.py
@@ -3,7 +3,7 @@ This module contains utility classes and functions for Roseau Load Flow.
 """
 from roseau.load_flow.utils.constants import CX, DELTA_P, EPSILON_0, EPSILON_R, MU_0, MU_R, OMEGA, PI, RHO, TAN_D, F
 from roseau.load_flow.utils.mixins import Identifiable, JsonMixin
-from roseau.load_flow.utils.types import BranchType, ConductorType, InsulationType, LineModel, LineType
+from roseau.load_flow.utils.types import ConductorType, InsulationType, LineModel, LineType
 
 __all__ = [
     # Constants
@@ -26,5 +26,4 @@ __all__ = [
     "LineModel",
     "ConductorType",
     "InsulationType",
-    "BranchType",
 ]